Unify propagator interface
