+++ src/geometry/LinearPropagator.hh
@@ -1,4 +1,4 @@
-//---------------------------------*-CUDA-*----------------------------------//
+//---------------------------------*-C++-*-----------------------------------//
 // Copyright 2020 UT-Battelle, LLC, and other Celeritas developers.
 // See the top-level COPYRIGHT file for details.
 // SPDX-License-Identifier: (Apache-2.0 OR MIT)
@@ -7,7 +7,6 @@
 //---------------------------------------------------------------------------//
 #pragma once
 
-#include "base/NumericLimits.hh"
 #include "GeoTrackView.hh"
 
 namespace celeritas
@@ -15,6 +14,8 @@ namespace celeritas
 //---------------------------------------------------------------------------//
 /*!
  * Propagate (move) a particle in a straight line.
+ *
+ * This can be called repeatedly until the track crosses a geometry boundary.
  */
 class LinearPropagator
 {
@@ -23,17 +24,17 @@ class LinearPropagator
     struct result_type
     {
         real_type distance; //!< Distance traveled
-        VolumeId  volume;   //!< Post-propagation volume
+        bool      boundary; //!< True if hit a boundary before given distance
     };
 
   public:
-    //! Construct from persistent and state data
+    // Construct from persistent and state data
     inline CELER_FUNCTION LinearPropagator(GeoTrackView* track);
 
-    //! Move track to next volume boundary.
+    // Move track to next volume boundary.
     inline CELER_FUNCTION result_type operator()();
 
-    //! Move track by a user-provided distance, or to the next boundary
+    // Move track by a user-provided distance, or to the next boundary
     inline CELER_FUNCTION result_type operator()(real_type dist);
 
   private:
@@ -41,6 +42,48 @@ class LinearPropagator
 };
 
 //---------------------------------------------------------------------------//
-} // namespace celeritas
+/*!
+ * Construct from persistent and state data.
+ */
+CELER_FUNCTION LinearPropagator::LinearPropagator(GeoTrackView* track)
+    : track_(*track)
+{
+    CELER_EXPECT(track);
+}
 
-#include "LinearPropagator.i.hh"
+//---------------------------------------------------------------------------//
+/*!
+ * Move track to next volume boundary.
+ */
+CELER_FUNCTION
+LinearPropagator::result_type LinearPropagator::operator()()
+{
+    result_type result;
+    result.distance = track_.move_to_boundary();
+    result.boundary = true;
+    return result;
+}
+
+//---------------------------------------------------------------------------//
+/*!
+ * Move track by a user-provided distance up to the next boundary.
+ */
+CELER_FUNCTION
+LinearPropagator::result_type LinearPropagator::operator()(real_type dist)
+{
+    CELER_EXPECT(dist > 0);
+
+    if (dist >= track_.next_step())
+    {
+        // Move to boundary
+        return (*this)();
+    }
+
+    result_type result;
+    result.distance = track_.move_internal(dist);
+    result.boundary = false;
+    return resuult;
+}
+
+//---------------------------------------------------------------------------//
+} // namespace celeritas
